requirements-cfd.txt
# 来自 cfd.yml 的 pip 部分，环境建好后安装：pip install -r requirements-cfd.txt
# jax/jaxlib 从 JAX 归档拉取（PyPI 已移除旧版）。CPU 即可。
--find-links https://storage.googleapis.com/jax-releases/jax_releases.html
absl-py==0.12.0
astunparse==1.6.3
attrs==21.2.0
cachetools==4.2.4
charset-normalizer==2.0.9
chex==0.1.0
cloudpickle==2.0.0
cycler==0.11.0
dask==2021.12.0
dill==0.3.4
dm-haiku==0.0.5
dm-tree==0.1.6
einops==0.3.2
execnet==1.9.0
flatbuffers==1.12
flax==0.3.6
fonttools==4.28.3
future==0.18.2
gast==0.4.0
gin-config==0.5.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.54.0
grpcio==1.34.1
h5py==3.1.0
idna==3.3
imageio==2.13.3
importlib-metadata==4.8.2
importlib-resources==5.4.0
iniconfig==1.1.1
jax==0.2.27
jaxlib==0.1.76
jmp==0.0.2
keras-nightly==2.5.0.dev2021032900
keras-preprocessing==1.1.2
kiwisolver==1.3.2
locket==0.2.1
markdown==3.3.6
matplotlib==3.5.0
msgpack==1.0.3
networkx==2.6.3
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
optax==0.1.0
packaging==21.3
pandas==1.3.4
partd==1.2.0
pillow==8.4.0
pluggy==1.0.0
promise==2.3
protobuf==3.19.1
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.6
pytest==6.2.5
pytest-forked==1.4.0
pytest-xdist==2.5.0
python-dateutil==2.8.2
pytz==2021.3
pywavelets==1.2.0
pyyaml==6.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
scikit-image==0.19.0
scipy==1.7.3
seaborn==0.11.2
setuptools-scm==6.3.2
six==1.15.0
tabulate==0.8.9
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
termcolor==1.1.0
tifffile==2021.11.2
toml==0.10.2
tomli==1.2.2
toolz==0.11.2
torch==1.10.0
tqdm==4.62.3
urllib3==1.26.7
werkzeug==2.0.2
wrapt==1.12.1
xarray==0.20.2
zipp==3.6.0
# typing-extensions 不写死：optax 要 3.10，tensorflow 要 3.7.4，由上面装完后的版本为准。
# TensorFlow 与 optax 冲突，下面 4 个包在 pip install -r 之后单独装（见 setup_cfd_env.sh）：
# tensorflow==2.5.0 tensorflow-datasets==4.4.0 tensorflow-estimator==2.5.0 tensorflow-metadata==1.5.0
